Made a buffer.rs submodule
